Refactor main navigation
