+++ src/components/MainNavigation/MainNavigation.tsx
@@ -1,59 +1,32 @@
-import { Form, NavLink, useRouteLoaderData } from "react-router-dom";
+import { useRouteLoaderData } from "react-router-dom";
 import "./MainNavigation.scss";
 import { getIsAdmin, getAuthUserId } from "../../utils/auth";
 
-import {
-  Box,
-  IconButton,
-  useTheme,
-  Typography,
-  useMediaQuery,
-  Avatar,
-} from "@mui/material";
-import { ReactElement, useContext } from "react";
-import { ColorModeContext, tokens } from "../../context/theme";
-import LightModeOutlinedIcon from "@mui/icons-material/LightModeOutlined";
-import DarkModeOutlinedIcon from "@mui/icons-material/DarkModeOutlined";
-import CalendarTodayOutlinedIcon from "@mui/icons-material/CalendarTodayOutlined";
+import { Box, useTheme, Typography, useMediaQuery } from "@mui/material";
+import { tokens } from "../../context/theme";
 import MenuOutlinedIcon from "@mui/icons-material/MenuOutlined";
-import PersonOutlinedIcon from "@mui/icons-material/PersonOutlined";
-import PeopleOutlineIcon from "@mui/icons-material/PeopleOutline";
+
 import HomeOutlinedIcon from "@mui/icons-material/HomeOutlined";
-import PersonAddAltIcon from "@mui/icons-material/PersonAddAlt";
-import LogoutIcon from "@mui/icons-material/Logout";
-import LoginIcon from "@mui/icons-material/Login";
-import AccountCircleIcon from "@mui/icons-material/AccountCircle";
-import logo from "../../assets/beawake-logo.png";
 
 import { useState } from "react";
 import { Sidebar, Menu, MenuItem } from "react-pro-sidebar";
 import { useQuery } from "@tanstack/react-query";
 import { getUserById } from "../../utils/http";
 import User from "../../models/user";
-import { apiUrl } from "../../utils/helpers";
+import { Token } from "../../utils/helpers";
+import UserProfile from "./UserProfile";
+import HeaderMenu from "./HeaderMenu";
+import AdminLinks from "./AdminLinks";
+import AuthLinks from "./AuthLinks";
+import CustomLink from "./CustomLink";
 
 const getIsLoggedAndNotExpired = (token: Token) => {
   return !!token;
 };
 
-type UserType = "admin" | "seller" | "hairsalon";
-
-const userTypeTranslations: Record<UserType, string> = {
-  admin: "Administrador",
-  seller: "Vendedor",
-  hairsalon: "Peluqueria",
-};
-
-const getUserType = (userType: UserType) => {
-  return userTypeTranslations[userType];
-};
-
-type Token = string | null;
-
 const MainNavigation: React.FC = () => {
   const theme = useTheme();
   const colors = tokens(theme.palette.mode);
-  const colorMode = useContext(ColorModeContext);
   const isNonMobile = useMediaQuery("(min-width: 600px)");
   const [isCollapsed, setIsCollapsed] = useState(!isNonMobile);
 
@@ -91,76 +64,14 @@ const MainNavigation: React.FC = () => {
             disabled={!isNonMobile}
           >
             {!isCollapsed && (
-              <Box
-                display="flex"
-                justifyContent="space-between"
-                alignItems="center"
-                ml="15px"
-              >
-                <img
-                  src={logo}
-                  alt="logo"
-                  className="logo"
-                  style={{ height: "28px" }}
-                />
-                <IconButton
-                  onClick={() => setIsCollapsed(!isCollapsed)}
-                  style={{ color: "white" }}
-                >
-                  <MenuOutlinedIcon />
-                </IconButton>
-              </Box>
+              <HeaderMenu
+                handleCollapsed={() => setIsCollapsed(!isCollapsed)}
+              />
             )}
           </MenuItem>
           {!isCollapsed && (
             <Box mb="25px">
-              <Box display="flex" justifyContent="center" alignItems="center">
-                {user?.avatar ? (
-                  <img
-                    alt="profile-user"
-                    width="100px"
-                    height="100px"
-                    src={user.avatar}
-                    style={{ cursor: "pointer", borderRadius: "50%" }}
-                  />
-                ) : (
-                  <Avatar sx={{ width: "100px", height: "100px" }}>
-                    <AccountCircleIcon
-                      sx={{
-                        width: "100%",
-                        height: "100%",
-                        bgcolor: colors.primary[200],
-                      }}
-                    />
-                  </Avatar>
-                )}
-              </Box>
-              <Box textAlign="center">
-                {isLoggedNotExpired && (
-                  <>
-                    <Typography
-                      variant="h5"
-                      fontWeight="bold"
-                      sx={{ m: "10px 0 0 0" }}
-                    >
-                      {userEmail}
-                    </Typography>
-                    <Typography variant="h5">
-                      {getUserType(user?.userType)}
-                    </Typography>
-                  </>
-                )}
-                <IconButton
-                  onClick={colorMode.toggleColorMode}
-                  style={{ color: "white" }}
-                >
-                  {theme.palette.mode === "dark" ? (
-                    <DarkModeOutlinedIcon />
-                  ) : (
-                    <LightModeOutlinedIcon />
-                  )}
-                </IconButton>
-              </Box>
+              <UserProfile userEmail={userEmail} user={user} />
             </Box>
           )}
           <Box>
@@ -171,95 +82,15 @@ const MainNavigation: React.FC = () => {
                 Secciones
               </Typography>
             )}
-            {isAdmin && (
-              <>
-                <CustomLink
-                  to="/usuarios"
-                  title="Usuarios"
-                  icon={<PeopleOutlineIcon />}
-                />
-                <CustomLink
-                  to="/profesionales"
-                  title="Profesionales"
-                  icon={<PeopleOutlineIcon />}
-                />
-              </>
-            )}
-            {!isLoggedNotExpired ? (
-              <>
-                <CustomLink
-                  to={`${apiUrl}/auth/google`}
-                  title="Iniciar Sesion"
-                  icon={<LoginIcon />}
-                />
-                <CustomLink
-                  to="/crear-usuario"
-                  title="Crear Usuario"
-                  icon={<PersonAddAltIcon />}
-                />
-              </>
-            ) : (
-              <>
-                <CustomLink
-                  to={`/usuarios/editar/${userId}`}
-                  title="Mi Perfil"
-                  icon={<PersonOutlinedIcon />}
-                />
-                <CustomLink
-                  to="/agenda"
-                  title="Agenda"
-                  icon={<CalendarTodayOutlinedIcon />}
-                />
-                <Form action="/logout" method="POST">
-                  <li className="ps-menuitem-root">
-                    <span className="ps-menu-button">
-                      <button type="submit" className="ps-menu-icon">
-                        <LogoutIcon />
-                      </button>
-
-                      <button type="submit" className="ps-menu-label">
-                        Salir
-                      </button>
-                    </span>
-                  </li>
-                </Form>
-              </>
-            )}
+            {isAdmin && <AdminLinks />}
+            <AuthLinks
+              isLoggedNotExpired={isLoggedNotExpired}
+              userId={userId}
+            />
           </Box>
         </Menu>
       </Sidebar>
     </Box>
-  );
-};
-
-const CustomLink = ({
-  to,
-  title,
-  icon,
-  ...props
-}: {
-  to: string;
-  title: string;
-  icon: ReactElement;
-}) => {
-  return (
-    <li className="ps-menuitem-root">
-      <NavLink
-        to={to}
-        {...props}
-        className={({ isActive }) =>
-          isActive ? "ps-menu-button ps-active" : "ps-menu-button"
-        }
-        end
-      >
-        <span className="ps-menu-icon">
-          <svg width="24" height="24">
-            {icon}
-          </svg>{" "}
-        </span>
-        <span className="ps-menu-label">{title}</span>
-      </NavLink>
-    </li>
   );
 };
